refactor(gateway): split session handlers
